Updated tests for sync catalog to algolia
